Send push notifications for unseen matches
