+++ src/notifications.js
@@ -47,3 +47,24 @@ export async function showLocalNotification(title, body) {
   }
   addNotification({ title, body, type: 'local' });
 }
+
+export async function sendPushNotification(body, userId) {
+  const base = process.env.FUNCTIONS_BASE_URL || '';
+  const send = async endpoint => {
+    const resp = await fetch(`${base}/.netlify/functions/${endpoint}`, {
+      method: 'POST',
+      headers: { 'Content-Type': 'application/json' },
+      body: JSON.stringify(userId ? { body, userId } : { body })
+    });
+    if (!resp.ok) {
+      const text = await resp.text();
+      console.error('Push failed', endpoint, text);
+    }
+  };
+  try {
+    await send('send-push');
+    await send('send-webpush');
+  } catch (err) {
+    console.error('Failed to send push', err);
+  }
+}
